Move output formatters into separate folder
