+++ src/server/plugins/engine/models/FormModel.ts
@@ -27,7 +27,6 @@ import {
 } from '~/src/server/plugins/engine/pageControllers/helpers.js'
 import { validationOptions as opts } from '~/src/server/plugins/engine/pageControllers/validationOptions.js'
 import * as defaultServices from '~/src/server/plugins/engine/services/index.js'
-import * as notifyServiceForMachines from '~/src/server/plugins/engine/services/notifyServiceForMachines.js'
 import {
   type FormContext,
   type FormContextRequest,
@@ -60,12 +59,6 @@ export class FormModel {
     services: Services = defaultServices
   ) {
     const result = formDefinitionSchema.validate(def, { abortEarly: false })
-
-    const servicesNew = {
-      // TODO do this conditionally
-      ...services,
-      outputService: notifyServiceForMachines
-    }
 
     if (result.error) {
       throw result.error
@@ -100,7 +93,7 @@ export class FormModel {
     this.values = result.value
     this.basePath = options.basePath
     this.conditions = {}
-    this.services = servicesNew
+    this.services = services
 
     def.conditions.forEach((conditionDef) => {
       const condition = this.makeCondition(conditionDef)
